Move flash_mla.h to kernels/params.h
