+++ csrc/kernels/get_mla_metadata.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include "flash_mla.h"
+#include "params.h"
 
 void run_get_mla_metadata_kernel(Mla_metadata_params &params, cudaStream_t stream);
