+++ csrc/kernels/mla_combine.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "flash_mla.h"
+#include "params.h"
 
 template<typename ElementT>
 void run_flash_mla_combine_kernel(Flash_fwd_mla_params &params, cudaStream_t stream);
