+++ csrc/kernels/splitkv_mla.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "flash_mla.h"
+#include "params.h"
 
 template<typename InputT>
 void run_flash_splitkv_mla_kernel(Flash_fwd_mla_params &params, cudaStream_t stream);
